requirements.txt
asgiref==3.4.1
certifi==2020.12.5
click==8.0.1
colorama==0.4.4
django-filter==2.4.0
djangorestframework==3.12.4
h11==0.12.0
importlib-metadata==4.8.1
Markdown==3.3.4
Pillow==8.2.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
wincertstore==0.2
zipp==3.5.0
